requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.2
cffi==1.17.1
channels==4.1.0
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==5.1.1
django-cors-headers==3.13.0
django-prose-editor==0.8.1
django-ratelimit==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.9
incremental==24.7.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
python-slugify==8.0.4
pytz==2022.7
service-identity==24.1.0
soupsieve==2.4.1
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
validate-email==1.3
zope.interface==7.0.3